Expose marquee support and honor pixel brightness
Add support for marquee text by exposing the marquee helper. Also
transform pixel hex values according to the supplied brightness level.

Finally adjust where the generated build is placed, in lib/esm, and
default to `--watch` on the ts build npm script.
