+++ lib/esm/index.js
@@ -1,11 +1,13 @@
 import { jsx as _jsx } from "react/jsx-runtime";
 import { useEffect, useRef } from "react";
-import { createDisplayEngine } from "@bigdots-io/display-engine";
-export { solidColorMacro, textMacro, twinkleMacro, meteorShowerMacro, } from "@bigdots-io/display-engine";
-export function updateDot(element, { y, x, hex }) {
+import { colorLuminance, createDisplayEngine, } from "@bigdots-io/display-engine";
+export { solidColor, text, twinkle, meteorShower, marquee, } from "@bigdots-io/display-engine";
+export function updateDot(element, { y, x, hex, brightness }) {
     var el = element.querySelectorAll(`[data-coordinates='${y}:${x}']`);
     if (el.length > 0) {
-        el[0].style.background = hex === "#000000" ? `#444` : hex;
+        if (!hex)
+            return;
+        el[0].style.background = colorLuminance(hex, (brightness * 10) / 100 - 1);
     }
 }
 function Row({ y, children, opacity, }) {
