+++ package.json
@@ -2,16 +2,17 @@
   "name": "@bigdots-io/react-client",
   "version": "0.2.2",
   "dependencies": {
+    "@bigdots-io/display-engine": "bigdots-io/display-engine.git#v0.3.0",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
-    "typescript": "^5.2.2",
-    "@bigdots-io/display-engine": "bigdots-io/display-engine.git#v0.2.0"
+    "typescript": "^5.2.2"
   },
-  "main": "lib/index.js",
-  "types": "lib/index.d.ts",
+  "main": "lib/esm/index.js",
+  "types": "lib/esm/index.d.ts",
+  "type": "module",
   "scripts": {
-    "build": "tsc"
+    "build": "tsc --watch"
   }
 }
